Adds ai model argument with default
Also refactors AI New functionality into a separate file
