+++ pkg/cmd/pulumi/new.go
@@ -15,15 +15,10 @@
 package main
 
 import (
-	"bufio"
-	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io/fs"
-	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -42,7 +37,6 @@ import (
 	"github.com/pulumi/pulumi/pkg/v3/engine"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/apitype"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag/colors"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/env"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/config"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/plugin"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/tokens"
@@ -56,9 +50,6 @@ import (
 
 const (
 	brokenTemplateDescription = "(This template is currently broken)"
-	NoSelection               = "no"
-	YesSelection              = "yes"
-	RefineSelection           = "refine"
 )
 
 type promptForValueFunc func(yes bool, valueType string, defaultValue string, secret bool,
@@ -82,18 +73,8 @@ type newArgs struct {
 	listTemplates     bool
 	aiPrompt          string
 	aiLanguage        string
+	aiModel           pulumiAIModel
 	templateMode      bool
-}
-
-func deriveAIOrTemplate(args newArgs) string {
-	if args.aiPrompt != "" || args.aiLanguage != "" {
-		return "ai"
-	}
-	return "template"
-}
-
-func shouldPromptForAIOrTemplate(args newArgs, userBackend backend.Backend) bool {
-	return args.aiPrompt == "" && args.aiLanguage == "" && !args.templateMode && userBackend.Name() == "pulumi.com"
 }
 
 func runNew(ctx context.Context, args newArgs) error {
@@ -451,205 +432,6 @@ func useSpecifiedDir(dir string) (string, error) {
 	return cwd, nil
 }
 
-type AIPromptRequestBody struct {
-	Language       string `json:"language"`
-	Instructions   string `json:"instructions"`
-	Model          string `json:"model"`
-	ResponseMode   string `json:"responseMode"`
-	ConversationID string `json:"conversationId"`
-	ConnectionID   string `json:"connectionId"`
-}
-
-// Iteratively prompt the user for input, sending their input as a prompt tp Pulumi AI
-// Stream the response back to the console, and repeat until the user is done.
-func runAINew(
-	ctx context.Context,
-	args newArgs,
-	opts display.Options,
-	userName string,
-) (conversationURL string, err error) {
-	languageOptions := []string{
-		"TypeScript",
-		"JavaScript",
-		"Python",
-		"Go",
-		"C#",
-	}
-	if args.aiLanguage == "" {
-		if err = survey.AskOne(&survey.Select{
-			Message: "Please select a language for your project:",
-			Options: languageOptions,
-		}, &args.aiLanguage, surveyIcons(opts.Color)); err != nil {
-			return "", err
-		}
-	}
-	var continuePrompt string
-	var connectionID string
-	var conversationID string
-	for continuePrompt, conversationURL, conversationID, connectionID, err = runAINewPromptStep(
-		opts,
-		args.aiLanguage,
-		args.aiPrompt,
-		args.yes,
-		userName,
-		"",
-		"",
-		"",
-		"",
-	); continuePrompt == RefineSelection; continuePrompt,
-		conversationURL,
-		conversationID,
-		connectionID,
-		err = runAINewPromptStep(
-		opts,
-		args.aiLanguage,
-		args.aiPrompt,
-		args.yes,
-		userName,
-		continuePrompt,
-		connectionID,
-		conversationURL,
-		conversationID,
-	) {
-		if err != nil {
-			return "", err
-		}
-	}
-	return conversationURL, err
-}
-
-func sendPromptToPulumiAI(
-	promptMessage string,
-	conversationID string,
-	connectionID string,
-	userName string,
-	language string,
-) (string, string, string, error) {
-	pulumiAIURL := env.AIServiceEndpoint.Value()
-	if pulumiAIURL == "" {
-		pulumiAIURL = "https://www.pulumi.com/ai"
-	}
-	parsedURL, err := url.Parse(pulumiAIURL)
-	if err != nil {
-		return "", "", "", err
-	}
-	requestPath := parsedURL.JoinPath("api", "chat")
-	requestBody := AIPromptRequestBody{
-		Language:       language,
-		Instructions:   promptMessage,
-		Model:          "gpt-4-turbo",
-		ResponseMode:   "code",
-		ConversationID: conversationID,
-		ConnectionID:   connectionID,
-	}
-	marshalledBody, err := json.Marshal(requestBody)
-	if err != nil {
-		return "", "", "", err
-	}
-	userDataCookie := http.Cookie{Name: "pulumi_command_line_user_name", Value: userName}
-	request, err := http.NewRequest("POST", requestPath.String(), bytes.NewReader(marshalledBody))
-	if err != nil {
-		return "", "", "", err
-	}
-	request.AddCookie(&userDataCookie)
-	fmt.Println("Sending prompt to Pulumi AI...")
-	res, err := http.DefaultClient.Do(request)
-	if err != nil {
-		return "", "", "", err
-	}
-	defer res.Body.Close()
-	reader := bufio.NewReader(res.Body)
-	fmt.Println("Pulumi AI response:")
-	for {
-		chunk, err := reader.ReadByte()
-		if err != nil && err.Error() != "EOF" {
-			return "", "", "", err
-		}
-		fmt.Print(string(chunk))
-		if err != nil && err.Error() == "EOF" {
-			// Add a newline to ensure we don't overwrite the last line of the Pulumi AI response
-			fmt.Println()
-			break
-		}
-	}
-	conversationID = res.Header.Get("x-conversation-id")
-	connectionID = res.Header.Get("x-connection-id")
-	conversationURL := parsedURL.JoinPath("api", "project", url.PathEscape(fmt.Sprintf("%s.zip", conversationID))).String()
-	return conversationURL, connectionID, conversationID, nil
-}
-
-func runAINewPromptStep(
-	opts display.Options,
-	language string,
-	prompt string,
-	yes bool,
-	userName string,
-	currentContinueSelection string,
-	connectionID string,
-	conversationURL string,
-	conversationID string,
-) (continueSelection string,
-	conversationURLReturn string,
-	conversationIDReturn string,
-	connectionIDReturn string,
-	err error,
-) {
-	var promptMessage string
-	if prompt == "" || currentContinueSelection != "" {
-		if err := survey.AskOne(&survey.Input{
-			Message: "Please input your prompt here:\n",
-		}, &promptMessage, surveyIcons(opts.Color)); err != nil {
-			return "", "", "", "", err
-		}
-	} else {
-		promptMessage = prompt
-	}
-	conversationURLReturn, connectionIDReturn, conversationIDReturn, err = sendPromptToPulumiAI(
-		promptMessage,
-		conversationID,
-		connectionID,
-		userName,
-		language,
-	)
-	if err != nil {
-		return "", "", "", "", err
-	}
-	if connectionID != "" && connectionID != connectionIDReturn {
-		connectionIDReturn = connectionID
-		fmt.Println("Connection ID changed, please restart the prompt")
-	}
-	if conversationID != "" && conversationID != conversationIDReturn {
-		return "", "", "", "", fmt.Errorf("conversation id %s changed to %s", conversationID, conversationIDReturn)
-	}
-	continuePromptOptions := []string{
-		RefineSelection,
-		YesSelection,
-		NoSelection,
-	}
-	continuePromptOptionsDescriptions := map[string]string{
-		RefineSelection: "Write a prompt to further refine this program",
-		YesSelection:    "Use this program to create the project",
-		NoSelection:     "Abort the prompt and exit",
-	}
-	if !yes {
-		if err := survey.AskOne(&survey.Select{
-			Message: "Use this program as a template?",
-			Options: continuePromptOptions,
-			Description: func(opt string, _ int) string {
-				return continuePromptOptionsDescriptions[opt]
-			},
-		}, &continueSelection, surveyIcons(opts.Color)); err != nil {
-			return "", "", "", "", err
-		}
-		if continueSelection == NoSelection {
-			return "", "", "", "", errors.New("aborting prompt")
-		}
-	} else {
-		continueSelection = YesSelection
-	}
-	return continueSelection, conversationURLReturn, conversationIDReturn, connectionIDReturn, nil
-}
-
 // newNewCmd creates a New command with default dependencies.
 // Intentionally disabling here for cleaner err declaration/assignment.
 //
@@ -805,6 +587,9 @@ func newNewCmd() *cobra.Command {
 	cmd.PersistentFlags().BoolVarP(
 		&args.templateMode, "template-mode", "t", false,
 		"Run in template mode, which will skip prompting for AI or Template functionality",
+	)
+	cmd.PersistentFlags().Var(
+		&args.aiModel, "ai-model", "Model to use for Pulumi AI ",
 	)
 
 	return cmd
@@ -1112,35 +897,6 @@ func pythonCommands() []string {
 	commands = append(commands, "python -m pip install -r requirements.txt")
 
 	return commands
-}
-
-// Prompt the user to decide whether they'd like to enter an interactive AI prompt or use a template.
-func chooseWithAIOrTemplate(opts display.Options) (string, error) {
-	// Customize the prompt a little bit (and disable color since it doesn't match our scheme).
-	surveycore.DisableColor = true
-
-	options := []string{
-		"ai",
-		"template",
-	}
-
-	optionsDescriptionMap := map[string]string{
-		"ai":       "Create a new Pulumi project using Pulumi AI",
-		"template": "Create a new Pulumi project using a template",
-	}
-
-	var ai string
-	if err := survey.AskOne(&survey.Select{
-		Message: "Would you like to create a new project with Pulumi AI?",
-		Options: options,
-		Description: func(opt string, _ int) string {
-			return optionsDescriptionMap[opt]
-		},
-	}, &ai, surveyIcons(opts.Color)); err != nil {
-		return "template", err
-	}
-
-	return ai, nil
 }
 
 // chooseTemplate will prompt the user to choose amongst the available templates.
